Small optimisation for building large polymers and placing solvents.