@ricalessandri 2024-08-13
@fgrunewald shall I do a small benchmark with the main branch and this branch? I could test a polymer melt and a single-chain simulation, tomorrow.