@fgrunewald 2024-08-13
@ricalessandri that would be fantastic; I briefly checked on Mac and should be fine. When you go for the melt you need to provide a bending option via the build file. I recommend trying a long chain perhaps 1500 residues. That is what one part of the speed-opt addresses. For PS you could use:

```
[ bending ]
PS PS PS 3
```

Using this bending setup we actually can generate a correct number of entanglements. I've done some benchmarking on the melt structures in the past week. Overall, I'm surprised how good they are if you provide a bending constant. Hopefully, I manage to share the details in Denver. 

The other part addresses polymer solutions. If you build a box of just water or polymer solvent mixture it should be significantly faster now and more linear in scaling.