@ricalessandri 2024-08-15
@fgrunewald I've tested the melt, making a fresh install just in case:
```
conda create -n polyply-speedopt python=3
conda activate polyply-speedopt
conda install -c conda numpy
conda install -c conda scipy
conda install -c conda numba
git clone git@github.com:marrink-lab/polyply_1.0.git polyply_1.0-speedopt
cd polyply_1.0-speedopt
pip install -e .
```
Then:
```
dens="1400"

cd polyply_1.0-speedopt
git checkout speedopt
echo "Polyply version: $(python -c "import polyply; print(polyply.__version__)")"   >  ../test_log_${dens}.txt
echo "Current Git commit: $(git rev-parse HEAD)"                                    >> ../test_log_${dens}.txt
echo "Current Git branch: $(git rev-parse --abbrev-ref HEAD)"                       >> ../test_log_${dens}.txt
cd ..

# PS
{ time polyply gen_coords -p system_melt.top -o test_melt.gro -name test_melt -dens ${dens} -b build_file.bld; } 2>> test_log_${dens}.txt

cd polyply_1.0-speedopt
git checkout master
echo "Polyply version: $(python -c "import polyply; print(polyply.__version__)")"   >> ../test_log_${dens}.txt
echo "Current Git commit: $(git rev-parse HEAD)"                                    >> ../test_log_${dens}.txt
echo "Current Git branch: $(git rev-parse --abbrev-ref HEAD)"                       >> ../test_log_${dens}.txt
cd ..

# PS
{ time polyply gen_coords -p system_melt.top -o test_melt.gro -name test_melt -dens ${dens} -b build_file.bld; } 2>> test_log_${dens}.txt
```

Which gives:
```
Polyply version: 1.7.1.dev47
Current Git commit: 6f168af11457ee68978110fad99b835661e6fc36
Current Git branch: speedopt
INFO - step - reading topology
INFO - step - processing topology
INFO - step - reading build options
INFO - step - generating templates
100%|██████████| 1000/1000 [00:02<00:00, 403.64it/s]
INFO - step - annotating ligands
INFO - step - generating system coordinates
100%|██████████| 1000/1000 [01:06<00:00, 14.98it/s]
INFO - step - backmapping to target resolution
100%|██████████| 1000/1000 [00:38<00:00, 26.02it/s]
INFO - step - writing output

real	2m5.287s
user	2m1.618s
sys	0m3.044s
Polyply version: 1.7.1.dev47
Current Git commit: c820aaa5e6da7bd8340b360a2a4453fca4d52138
Current Git branch: master
INFO - step - reading topology
INFO - step - processing topology
INFO - step - reading build options
INFO - step - generating templates
100%|██████████| 1000/1000 [00:02<00:00, 400.26it/s]
INFO - step - annotating ligands
INFO - step - generating system coordinates
100%|██████████| 1000/1000 [01:07<00:00, 14.91it/s]
INFO - step - backmapping to target resolution
100%|██████████| 1000/1000 [00:39<00:00, 25.13it/s]
INFO - step - writing output

real	2m6.855s
user	2m3.037s
sys	0m2.980s
```

I don't see much of a speed-up, but maybe the system needs to be even larger?
This is the system (gives a ~25^3 nm^3):
```
; Martini 3 particle definitions and interaction matrix
#include "/Users/alessandri/ric.alessandri_v2/Repos/Martini/itps/martini_v300/martini_v3.0.0.itp"
; Topologies of molecules in the system
#include "PS_n1500.itp"
[ system ]
PS melt
[ molecules ]
PS      1000
```