@fgrunewald 2024-08-15
@ricalessandri are you sure your polymer is 1500 units? 25^3nm^3 seems very small for that many residues?