@ricalessandri 2024-08-16
@fgrunewald good catch, I had indeed a typo in my itp generation script. I've repeated the test with an actual M3 PS with 1500 residues. For a system with 1000x 1500-mers (the box is ~63^3 nm^3), **I do find a speed-up of ~30%** (!), see below:
```
Polyply version: 1.7.1.dev47
Current Git commit: 6f168af11457ee68978110fad99b835661e6fc36
Current Git branch: speedopt
INFO - step - reading topology
INFO - step - processing topology
INFO - step - reading build options
INFO - step - generating templates
100%|██████████| 1000/1000 [00:35<00:00, 28.08it/s]
INFO - step - annotating ligands
INFO - step - generating system coordinates
100%|██████████| 1000/1000 [18:29<00:00,  1.11s/it]
INFO - step - backmapping to target resolution
100%|██████████| 1000/1000 [09:24<00:00,  1.77it/s]
INFO - step - writing output

real	33m36.302s
user	32m5.119s
sys	0m57.532s
Polyply version: 1.7.1.dev47
Current Git commit: c820aaa5e6da7bd8340b360a2a4453fca4d52138
Current Git branch: master
INFO - step - reading topology
INFO - step - processing topology
INFO - step - reading build options
INFO - step - generating templates
100%|██████████| 1000/1000 [00:35<00:00, 27.97it/s]
INFO - step - annotating ligands
INFO - step - generating system coordinates
100%|██████████| 1000/1000 [18:07<00:00,  1.09s/it]
INFO - step - backmapping to target resolution
100%|██████████| 1000/1000 [09:47<00:00,  1.70it/s]
INFO - step - writing output

real	48m31.518s
user	32m14.350s
sys	0m43.759s
```