@fgrunewald 2024-09-05
@pckroon they are still part of the nonbond_matrix; just the partitioning of the trees is differently